requirements.txt
pandas
numpy
scikit-learn
gradio
plotly
python-dotenv
pyarrow
SQLAlchemy
psycopg2-binary
